Global cleanup of imports and removal of warnings

+++ org.hpccsystems.eclide/src/org/hpccsystems/internal/data/HashCodeUtil.java
@@ -32,7 +32,7 @@ public final class HashCodeUtil {
 	 * chars.
 	 */
 	public static int hash( int aSeed, char aChar ) {
-		return firstTerm( aSeed ) + (int)aChar;
+		return firstTerm( aSeed ) + aChar;
 	}
 
 	/**
